Refactor zipDatadir
