+++ app/server/src/main/scala/org/bitcoins/server/BitcoinSAppConfig.scala
@@ -5,11 +5,9 @@ import com.typesafe.config.{Config, ConfigFactory}
 import grizzled.slf4j.Logging
 import org.bitcoins.chain.config.ChainAppConfig
 import org.bitcoins.commons.config.{AppConfig, ConfigOps}
-import org.bitcoins.commons.file.FileUtil
 import org.bitcoins.commons.util.ServerArgParser
 import org.bitcoins.core.config.NetworkParameters
 import org.bitcoins.core.util.{StartStopAsync, TimeUtil}
-import org.bitcoins.db.SQLiteUtil
 import org.bitcoins.dlc.node.config.DLCNodeAppConfig
 import org.bitcoins.dlc.wallet.DLCAppConfig
 import org.bitcoins.keymanager.config.KeyManagerAppConfig
@@ -18,12 +16,10 @@ import org.bitcoins.rpc.config.BitcoindRpcAppConfig
 import org.bitcoins.tor.config.TorAppConfig
 import org.bitcoins.wallet.config.WalletAppConfig
 
-import java.io.IOException
 import java.nio.file.{Files, Path, Paths}
 import java.util.concurrent.TimeUnit
 import scala.concurrent.Future
 import scala.concurrent.duration.{DurationInt, FiniteDuration}
-import scala.util.Try
 
 /** A unified config class for all submodules of Bitcoin-S
   * that accepts configuration. Thanks to implicit definitions
@@ -61,6 +57,8 @@ case class BitcoinSAppConfig(
     BitcoindRpcAppConfig(directory, confs: _*)
 
   lazy val network: NetworkParameters = chainConf.network
+
+  lazy val datadir: Path = directory
 
   /** Initializes the wallet, node and chain projects */
   override def start(): Future[Unit] = {
@@ -156,11 +154,6 @@ case class BitcoinSAppConfig(
   def withOverrides(configs: Config*): BitcoinSAppConfig = {
     BitcoinSAppConfig(directory, configs ++ confs: _*)
   }
-
-  /** Zips $HOME/.bitcoin-s
-    */
-  def zipDatadir(target: Path): Try[Path] =
-    BitcoinSAppConfig.zipDatadir(directory, target)
 }
 
 /** Implicit conversions that allow a unified configuration
@@ -261,45 +254,5 @@ object BitcoinSAppConfig extends Logging {
   /** Converts the given config to a bitcoind rpc config */
   def toBitcoindRpcConf(conf: BitcoinSAppConfig): BitcoindRpcAppConfig = {
     conf.bitcoindRpcConf
-  }
-
-  def zipDatadir(source: Path, target: Path): Try[Path] = Try {
-    if (Files.exists(target)) {
-      throw new IOException(
-        s"Cannot zip datadir. Target file already exists: $target")
-    }
-    val temp = Files.createTempDirectory(source, "backup")
-    try {
-      // we don't want to store chaindb.sqlite as these databases are huge
-      // skip logs and binaries as these can be large as well
-      val tempRE = (".*/" + temp.getFileName + "/.*").r
-
-      FileUtil.copyDirectory(
-        source = source,
-        target = temp,
-        fileNameFilter = Vector(".*.sqlite$".r,
-                                ".*.sqlite-shm$".r,
-                                ".*.sqlite-wal$".r,
-                                ".*bitcoin-s.log$".r,
-                                ".*/seeds/.*".r,
-                                ".*/tor/.*".r,
-                                ".*/binaries/.*".r,
-                                ".*.zip$".r,
-                                tempRE)
-      )
-
-      SQLiteUtil.backupDirectory(source = source,
-                                 target = temp,
-                                 fileNameFilter =
-                                   Vector(".*chaindb.sqlite$".r, tempRE))
-
-      FileUtil.zipDirectory(
-        source = temp,
-        target = target
-      )
-    } finally {
-      FileUtil.removeDirectory(temp)
-      ()
-    }
   }
 }
