+++ app/server/src/main/scala/org/bitcoins/server/CoreRoutes.scala
@@ -19,7 +19,6 @@ import upickle.default._
 
 import scala.collection.mutable
 import scala.concurrent.Future
-import scala.util.{Failure, Success}
 
 case class CoreRoutes()(implicit system: ActorSystem, config: BitcoinSAppConfig)
     extends ServerRoute {
@@ -219,17 +218,6 @@ case class CoreRoutes()(implicit system: ActorSystem, config: BitcoinSAppConfig)
               "redeemScript" -> Str(spk.hex)
             )
             Server.httpSuccess(json)
-          }
-      }
-
-    case ServerCommand("zipdatadir", arr) =>
-      withValidServerCommand(ZipDataDir.fromJsArr(arr)) {
-        case ZipDataDir(path) =>
-          complete {
-            config.zipDatadir(path) match {
-              case Success(_)  => Server.httpSuccess(ujson.Null)
-              case Failure(ex) => Server.httpError(ex.getMessage)
-            }
           }
       }
   }
